Add support for the multi-threaded AES-CTR cipher
This patch provides the multi-threaded AES-CTR cipher. Essentially,
several buffers are created that are filled with pre-computed keystreads
generates by worker threads. These buffers are drained in order so as to
provide a keystream 100% compatible with non-threaded AES-CTR
implementations. Prior to this release there was a significant flaw in
the way in which this cipher was used in SSH (the security of this cipher
was not impacted) where forking the process to the background would cause
the threads to become divorced from the parent process. This would lead
to a failure in port forwarding applications and, starting with 6.1p1,
incompatibility with the sandbox used by the privsep process. Both
problems were resoloved by using the non-threaded AES-CTR cipher during
authentication and then switching to the MT-AES-CTR cipher immediately
after.

(cherry picked from commit 20a5c9961eb343d187e8a5c5b25adf32f0e53fb0)
Signed-off-by: Robin H. Johnson <robbat2@gentoo.org>
(cherry picked from commit 5e89e9f1c29c85cdf10a9b754affffd98b535051)